LPD-35723 This is a logical bug. AssetEntryLocalServiceImpl.deleteEntry() deletes the AssetEntry before calling SocialActivityManagerUtil.deleteActivities(entry), then execution gets to SocialActivityLocalServiceImpl.deleteActivities() -> SocialActivityCounterLocalServiceImpl.deleteActivityCounters() which does AssetEntryPersistence.fetchByC_C(classNameId, classPK), it will never see the AssetEntry since it is already deleted. Not sure what to do, since this bad logic goes back to 813b7624 more than 14 years ago. Keep this existing logic, but use a thread local to bypass the useless AssetEntry refetching.

+++ portal-impl/src/com/liferay/portlet/social/service/impl/SocialActivityLocalServiceImpl.java
@@ -28,6 +28,8 @@ import com.liferay.portal.kernel.service.persistence.UserPersistence;
 import com.liferay.portal.kernel.transaction.TransactionCommitCallbackUtil;
 import com.liferay.portal.kernel.util.BulkDeleteCacheThreadLocal;
 import com.liferay.portal.kernel.util.MapUtil;
+import com.liferay.portal.kernel.util.PortalUtil;
+import com.liferay.portlet.asset.util.DeletedAssetEntryThreadLocal;
 import com.liferay.portlet.social.service.base.SocialActivityLocalServiceBaseImpl;
 import com.liferay.portlet.social.util.SocialActivityHierarchyEntry;
 import com.liferay.portlet.social.util.SocialActivityHierarchyEntryThreadLocal;
@@ -1176,8 +1178,15 @@ public class SocialActivityLocalServiceImpl
 			}
 		}
 
-		_socialActivityCounterLocalService.deleteActivityCounters(
-			classNameId, classPK);
+		String className = PortalUtil.getClassName(classNameId);
+
+		if (!className.equals(User.class.getName()) &&
+			!DeletedAssetEntryThreadLocal.isDeletedAssetEntry(
+				classNameId, classPK)) {
+
+			_socialActivityCounterLocalService.deleteActivityCounters(
+				classNameId, classPK);
+		}
 	}
 
 	protected boolean isLogActivity(SocialActivity activity) {
